Moved objectives to a separate package
